twitcher url to phoenix push
