+++ magpie.env
@@ -5,3 +5,4 @@ ADMIN_PASSWORD=qwerty
 PHOENIX_USER=phoenix
 PHOENIX_PASSWORD=qwerty
 PHOENIX_PORT=8443
+TWITCHER_PROTECTED_PATH=/ows/proxy
